+++ CHANGELOG.md
@@ -38,7 +38,7 @@
 
 * Requires >= PHP 7.3
 * Requires >= PHPUnit 8.0
-* Deleted the Application and File libraries
+* Deleted the Application, Debug, Environments, Events, File, HTTP, and Routing libraries
 * Removed all methods that were deprecated in v1.1.0
 * Added `void` return type to all void methods
 * Added nullable  type to all nullable method return types and parameters